examples: example1: download_data.py: tweak and fix download script
Seeing how the "data" directory isn't used by anywhere else other than
formation_energy.jl, set the directory to be set up in the example1/
directory itself.

Previously, if the "data" directory wasn't already present, the script
wouldn't run correctly, and would throw a FileNotFound Error. Instead,
check if the "data" directory exists, and create the same if it doesn't
exist.

Use absolute paths everywhere, so one can run the script correctly
even if they're not in the example1/ directory.

Update the script so it doesn't break when used with pymatgen version 2022.0.4
(essentially accomodating a version bump for the package dependency).

Add the entire "data" directory to be ignored.

Signed-off-by: Anant Thazhemadam <anant.thazhemadam@gmail.com>