examples: example1: formation_energy.jl: fully fix up example1
Update the datadir to be the absolute path as updated in #7c31f11.

Previously, just building the model directly right out of the box would
result in an AssertionError, as the pooled_feature_length (= 40) would
be greater than atom_conv_feature_length (= 32).
Granted the AssertionError, may not be raised everytime (as it depends
on the data being read and property for which the model is being trained),
nonetheless, the it still needs fixing. Fix the AssertionError, by
passing pooled_feature_length = atom_conv_feature_length/2 to the model
being constructed.

Signed-off-by: Anant Thazhemadam <anant.thazhemadam@gmail.com>